Created utils, and documented some functions
