+++ lib/dtn/dtnMethods.js
@@ -43,7 +43,6 @@ const handleAAPForwarder = async (method, modelName, fileHistory, data) => {
 async function handleCreationMethods(actionType, localParams) {
   console.log("THIS IS", this, this.create, this["create"]);
   console.log("actionType", actionType, "localParams", localParams);
-  console.log("is iterable", typeof localParams[Symbol.iterator] === "function");
 
   // Perform the desired operation on the local database.
   const mongooseResponse = await this[actionType](...localParams);
